remove unecassery code
now that I know what was uneeded for Heroku, I want to try and remove it. If this PR will have a working demo, it means we can do without all of this.

notice that the `env.PORT` variable is checked in `config.js` already, so it was just a duplication.